requirements.txt
accelerate==0.21.0
aiohttp==3.8.4
aiosignal==1.3.1
anthropic==0.3.4
anyio==3.7.1
appdirs==1.4.4
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
bentoml==1.0.23
build==0.10.0
cattrs==23.1.2
certifi==2023.5.7
charset-normalizer==3.2.0
circus==0.18.0
clarifai==9.5.4
clarifai-grpc==9.6.0
click==8.1.5
click-option-group==0.5.6
cloudpickle==2.2.1
cohere==3.10.0
colorama==0.4.6
coloredlogs==15.0.1
contextlib2==21.6.0
dataclasses-json==0.5.9
datasets==2.13.1
deepmerge==1.1.0
Deprecated==1.2.14
dill==0.3.6
distro==1.8.0
filelock==3.12.2
filetype==1.2.0
frozenlist==1.4.0
fs==2.4.16
fsspec==2023.6.0
googleapis-common-protos==1.59.1
greenlet==2.0.2
grpcio==1.56.0
grpcio-health-checking==1.48.2
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.0.1
inflection==0.5.1
Jinja2==3.1.2
langchain==0.0.234
langsmith==0.0.5
manifest-ml==0.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.14
mypy-extensions==1.0.0
networkx==3.1
nlpcloud==1.1.43
numexpr==2.8.4
numpy==1.25.1
openai==0.27.8
openapi-schema-pydantic==1.2.4
openllm==0.1.20
openlm==0.0.5
opentelemetry-api==1.17.0
opentelemetry-instrumentation==0.38b0
opentelemetry-instrumentation-aiohttp-client==0.38b0
opentelemetry-instrumentation-asgi==0.38b0
opentelemetry-instrumentation-grpc==0.38b0
opentelemetry-sdk==1.17.0
opentelemetry-semantic-conventions==0.38b0
opentelemetry-util-http==0.38b0
optimum==1.9.1
orjson==3.9.2
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
Pillow==10.0.0
pip-requirements-parser==32.0.1
pip-tools==7.0.0
prometheus-client==0.17.1
protobuf==3.20.3
psutil==5.9.5
pyarrow==12.0.1
pydantic==1.10.11
Pygments==2.15.1
pynvml==11.5.0
pyparsing==3.1.0
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python-multipart==0.0.6
python-rapidjson==1.10
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
redis==4.6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
safetensors==0.3.1
schema==0.7.5
sentencepiece==0.1.99
simple-di==0.1.5
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
sqlitedict==2.1.0
starlette==0.28.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.2
tokenizers==0.13.3
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
transformers==4.30.2
tritonclient==2.34.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==1.26.16
uvicorn==0.23.0
watchfiles==0.19.0
wcwidth==0.2.6
wrapt==1.15.0
xxhash==3.2.0
yarl==1.9.2
zipp==3.16.2
